chore: Extend .gitignore
The repository’s .gitignore was automatically generated by GitHub
when the repo was created and has some limitations.

I’m often bothered by mysterious .lock files appearing in the tree,
so I needed to update the .gitignore.

I decided to use the Linux Kernel project’s .gitignore as a reference,
removed entries that shouldn’t be ignored for this project, and then
merged the result with the current .gitignore.

Signed-off-by: WangYuli <wangyl5933@chinaunicom.cn>